+++ CHANGELOG.md
@@ -25,6 +25,7 @@ changes. Best viewed [here](https://google-grain.readthedocs.io/en/latest/change
   * Adds profiling of multiprocess workers when using XProf profiler. To enable,
     set flag `grain_enable_multiprocess_worker_profiling=true` and add
     `"profile_subprocesses" = True` in advanced profiler options.
+  * Adds support for non-sequence meta-features in packing transformations.
 
 * Breaking changes:
   * Force index shuffle rounds to not be bigger than 1024.